Были добавлены комментарии типа // MARK: - ... . Модели были вынесены из MovieQuizViewController в отедльные файлы в папку Models.
