+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,8 +1,7 @@
-//
-//  QuizQuestion.swift
-//  MovieQuiz
-//
-//  Created by Alexandr Artemov (Mac Mini) on 09.09.2023.
-//
+import UIKit
 
-import Foundation
+struct QuizQuestion {
+    let image: String
+    let text: String
+    let correctAnswer: Bool
+}
