+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,8 +1,7 @@
-//
-//  QuizResultsViewModel.swift
-//  MovieQuiz
-//
-//  Created by Alexandr Artemov (Mac Mini) on 09.09.2023.
-//
+import UIKit
 
-import Foundation
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
