+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,8 +1,7 @@
-//
-//  QuizStepViewModel.swift
-//  MovieQuiz
-//
-//  Created by Alexandr Artemov (Mac Mini) on 09.09.2023.
-//
+import UIKit
 
-import Foundation
+struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
